common config: Device, BlockDevice
