+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
   "cli",
   "client",
   "server",
+  "common",
 ]
 
 [profile.dev]
