+++ cli/Cargo.toml
@@ -14,3 +14,4 @@ toml = "0.5"
 serde = "1.0"
 serde_derive = "1.0"
 propolis = { path = "../propolis" }
+common = { path = "../common" }
