+++ common/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "common"
+version = "0.1.0"
+edition = "2018"
+authors = ["James MacMahon <james@oxide.computer>"]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+propolis = { path = "../propolis" }
+serde = "1.0"
+serde_derive = "1.0"
+toml = "0.5"
